# Micro-optimize type check in GVMDependenciesNode.SearchDynamicDependencies