--- a/src/coreclr/tools/aot/ILCompiler.Compiler/Compiler/DependencyAnalysis/GVMDependenciesNode.cs
+++ b/src/coreclr/tools/aot/ILCompiler.Compiler/Compiler/DependencyAnalysis/GVMDependenciesNode.cs
@@ -70,26 +70,23 @@ namespace ILCompiler.DependencyAnalysis
             bool methodIsShared = _method.IsSharedByGenericInstantiations;
 
             TypeSystemContext context = _method.Context;
-            object lastUninterestingObject = null;
 
             for (int i = firstNode; i < markedNodes.Count; i++)
             {
                 DependencyNodeCore<NodeFactory> entry = markedNodes[i];
+                EETypeNode entryAsEETypeNode;
 
                 // This method is often called with a long list of ScannedMethodNode
-                // or MethodCodeNode nodes. We are not interested in those and they
-                // would be rejected by the `entry is EETypeNode` check below. However,
-                // that check has to walk the whole class hierarchy and can get rather
-                // expensive, so take a shortcut here as micro optimization.
-                if (lastUninterestingObject is not null && entry.GetType() == lastUninterestingObject.GetType())
+                // or MethodCodeNode nodes. We are not interested in those. In order
+                // to make the type check as cheap as possible we check for specific
+                // *sealed* types instead of doing `entry is EETypeNode` which has
+                // to walk the whole class hierarchy for the non matching nodes.
+                if (entry is ConstructedEETypeNode constructedEETypeNode)
+                    entryAsEETypeNode = constructedEETypeNode;
+                else if (entry is CanonicalEETypeNode canonicalEETypeNode)
+                    entryAsEETypeNode = canonicalEETypeNode;
+                else
                     continue;
-
-                EETypeNode entryAsEETypeNode = entry as EETypeNode;
-                if (entryAsEETypeNode == null)
-                {
-                    lastUninterestingObject = entry;
-                    continue;
-                }
 
                 TypeDesc potentialOverrideType = entryAsEETypeNode.Type;
                 if (!potentialOverrideType.IsDefType || potentialOverrideType.IsInterface)