When compiling a large project, there is a considerable amount of time spent in `GVMDependenciesNode.SearchDynamicDependencies` on a type check in a loop. The whole compilation takes 4m 9s with this patch, and 4m 31s without the patch (with other unrelated optimizations in ILC). Profiler shows slightly smaller improvement of about ~15s which is still quite measurable.

Profile before:
```
  100%   SearchDynamicDependencies  •  46,057 ms  •  ILCompiler.DependencyAnalysis.GVMDependenciesNode.SearchDynamicDependencies(List, Int32, NodeFactory)
    79.9%   IsInstanceOfClass  •  36,779 ms  •  System.Runtime.CompilerServices.CastHelpers.IsInstanceOfClass(Void*, Object)
    0.97%   GetTypeDefinition  •  446 ms  •  Internal.TypeSystem.InstantiatedType.GetTypeDefinition()
```

Profile after:
```
  100%   SearchDynamicDependencies  •  31,337 ms  •  ILCompiler.DependencyAnalysis.GVMDependenciesNode.SearchDynamicDependencies(List, Int32, NodeFactory)
  ► 1.64%   GetTypeDefinition  •  515 ms  •  Internal.TypeSystem.InstantiatedType.GetTypeDefinition()
    0.75%   IsInstanceOfClass  •  236 ms  •  System.Runtime.CompilerServices.CastHelpers.IsInstanceOfClass(Void*, Object)
```