@filipnavara 2024-06-08
Note that there's still significant bottleneck in the `.GetType()` call and its equality. It's possible that https://github.com/dotnet/runtime/issues/88631 would measurably help in this case.

Switching to unsafe comparing of method tables instead of `Type` yields another 6 second savings under profiler.