@jkotas 2024-06-08
> Switching to unsafe comparing of method tables

A clean safe version of that would be:
```
object lastUninterestingObject = markedNodes;

...

if (entry.GetType() == lastUninterestingObject.GetType())
...
```
`a.GetType() == b.GetType()` is pattern matched by the JIT to compare method tables.