@filipnavara 2024-06-08
> A clean safe version of that would be:

That performs slightly better:
```
  100%   SearchDynamicDependencies  •  27,363 ms  •  ILCompiler.DependencyAnalysis.GVMDependenciesNode.SearchDynamicDependencies(List, Int32, NodeFactory)
  ► 1.61%   GetTypeDefinition  •  441 ms  •  Internal.TypeSystem.InstantiatedType.GetTypeDefinition()
  ► 0.78%   FindVirtualFunctionTargetMethodOnObjectType  •  214 ms  •  Internal.TypeSystem.MetadataVirtualMethodAlgorithm.FindVirtualFunctionTargetMethodOnObjectType(MethodDesc, MetadataType)
    0.66%   IsInstanceOfClass  •  179 ms  •  System.Runtime.CompilerServices.CastHelpers.IsInstanceOfClass(Void*, Object)
```