@filipnavara 2024-06-08
> Avoiding null check should make it a bit faster

It's pretty much within the margin of error (0.5s - 0.8s) but it's slightly faster.