@MichalStrehovsky 2024-06-10
I wonder if it would be cleaner to just seal `ConstructedEETypeNode` and turn this into an `is ConstructedEETypeNode or CanonicalEETypeNode`.

Honestly, neither of these options are particularly nice. The PR proposal is hardcoding assumptions that there is something else in the graph that may not be interesting (a situation that only exists after #99037, a completely unrelated change - dynamic dependencies are not used in the compiler much - and the other spot using dynamic dependencies is not getting this optimization). The other option is assuming these descendants are sealed, but at least that structure has been constant for many years.