@MichalStrehovsky 2024-06-10
> > I wonder if it would be cleaner to just seal `ConstructedEETypeNode` and turn this into an `is ConstructedEETypeNode or CanonicalEETypeNode`.
> 
> Unfortunately, written like that, it would not solve the issue. The expensive part is when the check returns `false` because the whole type hierarchy is walked to the top. Of course it’s possible to write the check as direct type check but it feels just as ugly as this PR.

An `is` check with a sealed type is a simple MethodTable comparison with no inheritance chain walking. Or am I missing something?