@filipnavara 2024-06-10
> The PR proposal is hardcoding assumptions that there is something else in the graph that may not be interesting

Yes, but at least it’s not hard-coding what the uninteresting nodes are. In fact, we cannot easily hard code it anyway since `MethodWithCodeNode` is not accessible in the same assembly.

Ideally, we could rethink the whole concept and somehow filter it before it reaches this point but that would be very intrusive change.