@filipnavara 2024-06-10
> An `is` check with a sealed type is a simple MethodTable comparison with no inheritance chain walking.

Ah, you are probably right.