@filipnavara 2024-06-10
> I wonder if it would be cleaner to just seal `ConstructedEETypeNode` and turn this into an `is ConstructedEETypeNode or CanonicalEETypeNode`.

Done.

There separate profiles of this variation ended up in the 25s - 31s range. The profile breakdown suggests that it's getting us the same 30%+ time saving.